scripts/requirements.txt
Pillow==10.0.0
mysql-connector-python==8.1.0
requests==2.31.0